Day 14 Scope
Objective
Today we're discussing scope. Check out the Tutorial tab for learning materials and an instructional video!

The absolute difference between two integers,a  and b, is written as |a-b|. The maximum absolute difference between two integers in a set of positive integers, elements, is the largest absolute difference between any two integers in  elements.


The Difference class is started for you in the editor. It has a private integer array (elements) for storing N  non-negative integers, and a public integer (maximumDiffrence) for storing the maximum absolute difference.

Task
Complete the Difference class by writing the following:

A class constructor that takes an array of integers as a parameter and saves it to the elements  instance variable.
A computeDifference method that finds the maximum absolute difference between any 2  numbers in elements and stores it in the  maxmum difference instance variable.
Input Format

You are not responsible for reading any input from stdin. The locked Solution class in the editor reads in 2 lines of input. The first line contains N , the size of the elements array. The second line has N space-separated integers that describe the  elements array.